requirements.txt
fal-client==0.5.6
pydantic>=2.5.2,<3.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0 